requirements.txt
openai
pillow
opencv-python-headless
numpy